Dodanie kolorow

+++ src/ciastkarnia.c
@@ -1,6 +1,5 @@
 #include "ciastkarnia.h"
 
-
 void handle_error(const char *msg) {
     perror(msg);
     exit(EXIT_FAILURE);
@@ -8,11 +7,10 @@ void handle_error(const char *msg) {
 
 int sem_op(int semid, int sem_num, int op) {
     struct sembuf sb = {
-        .sem_num = sem_num,
+        .sem_num = (unsigned short)sem_num,
         .sem_op  = (short)op,
         .sem_flg = SEM_UNDO
     };
-
     while (semop(semid, &sb, 1) == -1) {
         if (errno == EINTR) continue;
         return -1;
